Enable create settings to access resource attributes
