+++ extension/extension_test.go
@@ -28,6 +28,7 @@ import (
 type nopExtension struct {
 	component.StartFunc
 	component.ShutdownFunc
+	CreateSettings
 }
 
 func TestNewFactory(t *testing.T) {
@@ -91,7 +92,6 @@ func TestMakeFactoryMap(t *testing.T) {
 func TestBuilder(t *testing.T) {
 	const typeStr = "test"
 	defaultCfg := struct{}{}
-	nopExtensionInstance := new(nopExtension)
 	testID := component.NewID(typeStr)
 	unknownID := component.NewID("unknown")
 
@@ -100,7 +100,7 @@ func TestBuilder(t *testing.T) {
 			typeStr,
 			func() component.Config { return &defaultCfg },
 			func(ctx context.Context, settings CreateSettings, extension component.Config) (Extension, error) {
-				return nopExtensionInstance, nil
+				return nopExtension{CreateSettings: settings}, nil
 			},
 			component.StabilityLevelDevelopment),
 	}...)
@@ -112,6 +112,11 @@ func TestBuilder(t *testing.T) {
 	e, err := b.Create(context.Background(), createSettings(testID))
 	assert.NoError(t, err)
 	assert.NotNil(t, e)
+
+	// Check that the extension has access to the resource attributes.
+	nop, ok := e.(nopExtension)
+	assert.True(t, ok)
+	assert.Len(t, nop.CreateSettings.ResourceAttrs, 0)
 
 	missingType, err := b.Create(context.Background(), createSettings(unknownID))
 	assert.EqualError(t, err, "extension factory not available for: \"unknown\"")
